fix(tests): setup not done properly
